backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
aiosqlite==0.20.0
aiohttp==3.10.0
onvif-zeep==0.2.12
pyyaml==6.0.2
python-multipart==0.0.9
